change all float values to double

+++ Kal/act3/src/com/kal/OvernightPackage.java
@@ -5,17 +5,17 @@ public class OvernightPackage extends Package{
     private int senderZip;
     private String recipientName, recipientAddress, recipientCity, recipientState;
     private int recipientZip;
-    private static float packWeight;
-    private static float packCost;
-    float totalCost;
-    private static float additionalFee;
-    public OvernightPackage(String senderName, String senderAddress, String senderCity, String senderState, int senderZip, String recipientName, String recipientAddress, String recipientCity, String recipientState, int recipientZip, float packageWeight, float packageCost, float additionalOvernight) {
+    private static double packWeight;
+    private static double packCost;
+    double totalCost;
+    private static double additionalFee;
+    public OvernightPackage(String senderName, String senderAddress, String senderCity, String senderState, int senderZip, String recipientName, String recipientAddress, String recipientCity, String recipientState, int recipientZip, double packageWeight, double packageCost, double additionalOvernight) {
         super(senderName, senderAddress, senderCity, senderState, senderZip, recipientName, recipientAddress, recipientCity, recipientState, recipientZip, packageWeight, packageCost);
         this.senderName = senderName; this.senderAddress = senderAddress; this.senderCity = senderCity; this.senderState = senderState; this.senderZip = senderZip;
         this.recipientName = recipientName; this.recipientAddress = recipientAddress; this.recipientCity = recipientCity; this.recipientState = recipientState; this.recipientZip = recipientZip;
         packCost = packageCost; packWeight = packageWeight; OvernightPackage.additionalFee = additionalOvernight; totalCost = totalCalcCost();
     }
-    public static float totalCalcCost(){
+    public static double totalCalcCost(){
         return (packCost+additionalFee)*packWeight;
     }
     public String toString(){
